Updated functions, fixed bugs

+++ lesson-9/FileManagerr/FileManagerr/Functions.cs
@@ -6,24 +6,28 @@ namespace FileManager
 {
     public static class Functions
     {
-        public static void Header()
+        /// <summary>
+        /// Метод построения верхней "шапки"
+        /// </summary>
+        private static void Header()
         {
-            Console.SetWindowSize(120, 30);
             Console.BackgroundColor = ConsoleColor.DarkBlue;
             Console.ForegroundColor = ConsoleColor.White;
-            Console.WriteLine("┌─────────────────────────────────────────────────────FILE-MANAGER───────────────────────────────────────────────────┐");
-            Console.WriteLine("│ COMMANDS:                                                                                                          │");
-            Console.WriteLine("│ ls <path> - Просмотр файлов каталога                             cp <path1> <path2> - Копирование файла (каталога) │");
-            Console.WriteLine("│                                                                                                                    │");
-            Console.WriteLine("│ dir <path> - Информация о катаголе                               clear - для очистки консоли                       │");
-            Console.WriteLine("│                                                                                                                    │");
-            Console.WriteLine("│ rm <path> - Удаление каталога(рекурсивно)/файла                  file <file> - Вывод информации о файле            │");
-            Console.WriteLine("│                                                                                                                    │");
-            Console.WriteLine("│ Введите quit, q, exit для выхода                                                                                   │");
-            Console.WriteLine("└────────────────────────────────────────────────────────────────────────────────────────────────────────────────────┘");
-            
+            Console.SetWindowSize(121, 30);
+            Console.WriteLine("┌─────────────────────────────────────────────────────FILE-MANAGER─────────────────────────────────────────────────────┐");
+            Console.WriteLine("│ COMMANDS:                                                                                                            │");
+            Console.WriteLine("│ ls <path> - Просмотр файлов каталога                                cp <path1> <path2> - Копирование файла (каталога)│");
+            Console.WriteLine("│                                                                                                                      │");
+            Console.WriteLine("│ dir <path> - Информация о катаголе                                  clear - для очистки консоли                      │");
+            Console.WriteLine("│                                                                                                                      │");
+            Console.WriteLine("│ rm <path> - Удаление каталога(рекурсивно)/файла                     file <file> - Вывод информации о файле           │");
+            Console.WriteLine("│                                                                                                                      │");
+            Console.WriteLine("│ Введите quit, q, exit для выхода                                                                                     │");
+            Console.WriteLine("└──────────────────────────────────────────────────────────────────────────────────────────────────────────────────────┘");
         }
-
+        /// <summary>
+        /// Метод сохранения конфигурационного файла "config.cfg"
+        /// </summary>
         public static void SerializeConf()
         {
             var filename = "config.cfg";
@@ -32,18 +36,32 @@ namespace FileManager
             var confString = new[] {"lastDirectory = " + lastDir,"pageCount = " + pageCount.ToString()};
             File.WriteAllLines(filename,confString);
         }
-
+        /// <summary>
+        /// Метод открытия конфигурационного файла "config.cfg" и применения изменения согласно файлу
+        /// </summary>
         public static void DeserializeConf()
         {
-            var filename = "config.cfg";
-            var confString = File.ReadAllLines(filename);
-            var lastDir = confString[0].Split(' ');
-            var pageCount = confString[1].Split(' ');
-            FileManagerr.Properties.Settings.Default.lastDirectory = lastDir[2];
-            FileManagerr.Properties.Settings.Default.pageCount = int.Parse(pageCount[2]);
+            try
+            {
+                var filename = "config.cfg";
+                var confString = File.ReadAllLines(filename);
+                var lastDir = confString[0].Split(' ');
+                var pageCount = confString[1].Split(' ');
+                FileManagerr.Properties.Settings.Default.lastDirectory = lastDir[2];
+                FileManagerr.Properties.Settings.Default.pageCount = int.Parse(pageCount[2]);
+            }
+            catch (FileNotFoundException) //обработка исключения отсутствия файла
+            {
+                SerializeConf();
+            }
+            
 
         }
-        
+        /// <summary>
+        /// Просмотр дерева каталогов
+        /// </summary>
+        /// <param name="currentDirectory"></param>
+        /// <param name="page"></param>
         public static void DirectoryViewer(string currentDirectory,int page)
         {
             Header();
@@ -69,7 +87,10 @@ namespace FileManager
                 Console.WriteLine("Недопустимые знаки.");
             }
         }
-
+        /// <summary>
+        /// Просмотр файлов каталога
+        /// </summary>
+        /// <param name="targetDirectory"></param>
         private static void FileViewer(string targetDirectory)
         { 
             try
@@ -86,11 +107,14 @@ namespace FileManager
             }
             
         }
-
+        /// <summary>
+        /// Обработчик консольных комманд
+        /// </summary>
+        /// <param name="command"></param>
         public static void CommandHandler(string command)
         {
             var switchArray = command.Split(' ');
-            if (switchArray[0].ToUpper() == "LS")
+            if (switchArray[0].ToUpper() == "LS") //Обработка команды перехода в директорию
             {
                 Console.Clear();
                 if (switchArray.Length < 2)
@@ -107,7 +131,7 @@ namespace FileManager
                     {
                         if (switchArray.Length >= 3)
                         {
-                            if (switchArray[2].ToUpper() == "-P")
+                            if (switchArray[2].ToUpper() == "-P") //Обработка параметра постраничного вывода
                             {
                                 try
                                 {
@@ -136,19 +160,19 @@ namespace FileManager
                 }
             }
 
-            if (switchArray[0].ToUpper() == "CP")
-            { 
-                Console.Clear();
+            if (switchArray[0].ToUpper() == "CP") //Обработка команды копирования файла/каталога
+            {
                 try
                 {
                     if (switchArray.Length == 3)
                     {
-                        if (Directory.Exists(switchArray[1]) && Directory.Exists(switchArray[2]))
+                        if (Directory.Exists(switchArray[1]) && Directory.Exists(switchArray[2])) //Проверка на наличие каталога, иначе вызов метода работы с файлом
                         {
                             CopyCatalog(switchArray[1], switchArray[2]);
                         }
                         else
                         {
+                            
                             CopyFile(switchArray[1], switchArray[2]);
                         }
                     }
@@ -167,7 +191,7 @@ namespace FileManager
                 }
             }
 
-            if (switchArray[0].ToUpper() == "FILE")
+            if (switchArray[0].ToUpper() == "FILE") //обработка команды вывода информации о файле
             {
                 try
                 {
@@ -179,7 +203,7 @@ namespace FileManager
                 }
             }
 
-            if (switchArray[0].ToUpper() == "DIR")
+            if (switchArray[0].ToUpper() == "DIR") //Обработка команды вывода информации о директории
             {
                 try
                 {
@@ -195,13 +219,13 @@ namespace FileManager
                 }
             }
 
-            if (switchArray[0].ToUpper() == "CLEAR")
+            if (switchArray[0].ToUpper() == "CLEAR") //Обработка команды очистки консоли
             {
                 Console.Clear();
                 Header();
             }
 
-            if (switchArray[0].ToUpper() == "RM")
+            if (switchArray[0].ToUpper() == "RM") // Обработка команды удаления
             {
                 if (File.Exists(switchArray[1]))
                 {
@@ -213,12 +237,21 @@ namespace FileManager
                 }
             }
         }
+        /// <summary>
+        /// Метод копирования файла
+        /// </summary>
+        /// <param name="firstFile"></param>
+        /// <param name="secondFile"></param>
         private static void CopyFile(string firstFile, string secondFile)
         {
             File.Copy(firstFile, secondFile);
             Console.WriteLine($"Файл {firstFile} успешно скопирован");
         }
-
+        /// <summary>
+        /// Метод копирования каталога
+        /// </summary>
+        /// <param name="firstCatalog"></param>
+        /// <param name="secondCatalog"></param>
         private static void CopyCatalog(string firstCatalog, string secondCatalog)
         {
             foreach (var directory in Directory.GetDirectories(firstCatalog,"*",SearchOption.AllDirectories))
@@ -231,7 +264,10 @@ namespace FileManager
                 File.Copy(file, file.Replace(firstCatalog, secondCatalog),true);
             }
         }
-
+        /// <summary>
+        /// Вывод информации о директории
+        /// </summary>
+        /// <param name="directoryPath"></param>
         private static void DirectoryInfo(string directoryPath)
         {
             try
@@ -258,25 +294,40 @@ namespace FileManager
                 Console.WriteLine("Имя папки не может быть пустое");
             }
         }
+        /// <summary>
+        /// Удаление директории (рекурсивно)
+        /// </summary>
+        /// <param name="directoryPath"></param>
         private static void DeleteDirectory(string directoryPath)
         {
-            try
+            if (Directory.Exists(directoryPath))
             {
-                Console.WriteLine($"Вы уверены что хотите удалить {directoryPath}? В ней могут находиться файлы Yes/No");
-                var choose = Console.ReadLine();
-                if (choose != null && (choose.ToUpper() == "Y" || choose.ToUpper() == "YES"))
-                { 
-                    Directory.Delete(directoryPath,true);
-                    Console.WriteLine($"Каталог {directoryPath} успешно удален!");
-                } else Console.WriteLine("Удаление каталога отменено.");
+                try
+                {
+                    Console.WriteLine($"Вы уверены что хотите удалить {directoryPath} Yes/No? В ней могут находиться файлы!");
+                    var choose = Console.ReadLine();
+                    if (choose != null && (choose.ToUpper() == "Y" || choose.ToUpper() == "YES"))
+                    { 
+                        Directory.Delete(directoryPath,true);
+                        Console.WriteLine($"Каталог {directoryPath} успешно удален!");
+                    } else Console.WriteLine("Удаление каталога отменено.");
                
+                }
+                catch (UnauthorizedAccessException)
+                {
+                    Console.WriteLine("Вы не можете удалить этот каталог"); 
+                } 
             }
-            catch (UnauthorizedAccessException)
+            else
             {
-                Console.WriteLine("Вы не можете удалить этот каталог");
+                Console.WriteLine("Такой директории не существует.");
             }
             
         }
+        /// <summary>
+        /// Удаление файла
+        /// </summary>
+        /// <param name="filename"></param>
         private static void DeleteFile(string filename)
         {
             try
@@ -296,7 +347,10 @@ namespace FileManager
                 Console.WriteLine("Вы не можете удалить этот файл");
             }
         }
-
+        /// <summary>
+        /// Вывод информации о файле
+        /// </summary>
+        /// <param name="filename"></param>
         private static void FileInformation(string filename)
         {
             try
